Move grow inside commands.
